requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
aiohttp==3.9.5
tldextract==5.1.2
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml[html_clean]==5.2.2
google-generativeai==0.8.3
keybert==0.8.5
sentence-transformers==3.0.1
yake==0.4.8
spacy==3.7.5
textblob==0.17.1
